+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/views_env_variables.cljs
@@ -1,4 +1,4 @@
-(ns sixsq.nuvla.ui.deployment-dialog.views-parameters
+(ns sixsq.nuvla.ui.deployment-dialog.views-env-variables
   (:require
     [re-frame.core :refer [dispatch subscribe]]
     [sixsq.nuvla.ui.deployment-dialog.events :as events]
@@ -12,48 +12,48 @@
 
 (defn summary-row
   []
-  (let [tr           (subscribe [::i18n-subs/tr])
-        input-params (subscribe [::subs/input-parameters])
-        completed?   (subscribe [::subs/parameters-completed?])
+  (let [tr            (subscribe [::i18n-subs/tr])
+        env-variables (subscribe [::subs/env-variables])
+        completed?    (subscribe [::subs/env-variables-completed?])
 
-        description  (str "Number of parameters: " (count @input-params))
-        on-click-fn  #(dispatch [::events/set-active-step :parameters])]
+        description   (str "Count: " (count @env-variables))
+        on-click-fn   #(dispatch [::events/set-active-step :environmental-variables])]
 
-    ^{:key "parameters"}
+    ^{:key "env-variables"}
     [ui/TableRow {:active   false
                   :on-click on-click-fn}
      [ui/TableCell {:collapsing true}
       (if @completed?
         [ui/Icon {:name "list alternate outline", :size "large"}]
         [ui/Icon {:name "warning sign", :size "large", :color "red"}])]
-     [ui/TableCell {:collapsing true} (@tr [:parameters])]
+     [ui/TableCell {:collapsing true} (@tr [:environmental-variables])]
      [ui/TableCell [:div [:span description]]]]))
 
 
 (defn as-form-input
-  [{:keys [parameter description value] :as param}]
+  [{env-name :name env-description :description env-value :value env-required :required :as env-variable}]
   (let [deployment (subscribe [::subs/deployment])]
-    ^{:key parameter}
-    [ui/FormField
-     [:label parameter ff/nbsp (ff/help-popup description)]
+    ^{:key env-name}
+    [ui/FormField {:required env-required}
+     [:label env-name ff/nbsp (ff/help-popup env-description)]
      [ui/Input
       {:type          "text"
-       :name          parameter
-       :default-value (or value "")
+       :name          env-name
+       :default-value (or env-value "")
        :read-only     false
        :fluid         true
        :on-blur       (ui-callback/input-callback
                         (fn [new-value]
-                          (let [updated-deployment (utils/update-parameter-in-deployment parameter new-value @deployment)]
+                          (let [updated-deployment (utils/update-env-variable-in-deployment env-name new-value @deployment)]
                             (dispatch [::events/set-deployment updated-deployment]))))}]]))
 
 
 (defn content
   []
-  (let [tr           (subscribe [::i18n-subs/tr])
-        input-params (subscribe [::subs/input-parameters])]
+  (let [tr            (subscribe [::i18n-subs/tr])
+        env-variables (subscribe [::subs/env-variables])]
 
-    (if (seq @input-params)
+    (if (seq @env-variables)
       (vec (concat [ui/Form]
-                   (map as-form-input @input-params)))
+                   (map as-form-input @env-variables)))
       [ui/Message {:success true} (@tr [:no-input-parameters])])))