+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/views_summary.cljs
@@ -5,7 +5,7 @@
     [sixsq.nuvla.ui.deployment-dialog.subs :as subs]
     [sixsq.nuvla.ui.deployment-dialog.views-credentials :as credentials-step]
     [sixsq.nuvla.ui.deployment-dialog.views-data :as data-step]
-    [sixsq.nuvla.ui.deployment-dialog.views-parameters :as parameters-step]
+    [sixsq.nuvla.ui.deployment-dialog.views-env-variables :as env-variables-step]
     [sixsq.nuvla.ui.i18n.subs :as i18n-subs]
     [sixsq.nuvla.ui.utils.semantic-ui :as ui]))
 
@@ -39,5 +39,5 @@
       (when @data-step-active?
         [data-step/summary-row])
       [credentials-step/summary-row]
-      [parameters-step/summary-row]
+      [env-variables-step/summary-row]
       ]]))